lambda_search/requirements.txt
boto3
psycopg2-binary
PyMuPDF
tiktoken